requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async HTTP client
aiohttp==3.9.1
aiofiles==23.2.1

# Web3 and Ethereum
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1
hexbytes==0.3.1

# Polymarket specific
py-clob-client==0.1.8  # Official Polymarket client (optional)

# Database (optional for persistence)
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utils
python-dateutil==2.8.2
pytz==2023.3